+++ AmazonChimeSDK/AmazonChimeSDK/ingestion/EventClientConfiguration.swift
@@ -25,6 +25,21 @@ import Foundation
 
 /// `EventClientConfiguration` contains speciic data required to send as metadata.
 @objc public protocol EventClientConfiguration {
+    
+    /**
+     The type of the Ingestion event
+     - Attention: replaced with `tag`
+     */
     var type: EventClientType { get }
     var eventClientJoinToken: String { get }
+    
+    /**
+     Tagging the source of the events, which will be translated to `Type` for Ingestion event
+     */
+    var tag: String { get }
+    
+    /**
+     The attributes that will be sent to Ingestion Service as metadata along with common attributes
+     */
+    var metadataAttributes: [String: Any] { get }
 }