+++ AmazonChimeSDK/AmazonChimeSDK/internal/ingestion/database/IngestionMeetingEvent.swift
@@ -13,5 +13,59 @@ import Foundation
 /// Swift doesn't decode Any type very well and it has to be one of pritimitve types.
 struct IngestionMeetingEvent: Codable {
     let name: String
-    let eventAttributes: IngestionEventAttributes
+    var eventAttributes: [String: AnyCodable?]
+}
+
+// MARK: Helper methods for retriving data from `eventAttributes`
+extension IngestionMeetingEvent {
+    
+    func getMeetingId() -> String? {
+        let item = eventAttributes[EventAttributeName.meetingId.description]
+        return item??.value as? String
+    }
+    
+    func getTimestampMs() -> Int64? {
+        let item = eventAttributes[EventAttributeName.timestampMs.description]
+        return item??.int64Value
+    }
+    
+    func getMaxVideoTileCount() -> Int? {
+        let item = eventAttributes[EventAttributeName.maxVideoTileCount.description]
+        return item??.value as? Int
+    }
+    
+    func getMeetingStartDurationMs() -> Int64? {
+        let item = eventAttributes[EventAttributeName.meetingStartDurationMs.description]
+        return item??.int64Value
+    }
+    
+    func getMeetingDurationMs() -> Int64? {
+        let item = eventAttributes[EventAttributeName.meetingDurationMs.description]
+        return item??.int64Value
+    }
+    
+    func getMeetingErrorMessage() -> String? {
+        let item = eventAttributes[EventAttributeName.meetingErrorMessage.description]
+        return item??.value as? String
+    }
+    
+    func getMeetingStatus() -> String? {
+        let item = eventAttributes[EventAttributeName.meetingStatus.description]
+        return item??.value as? String
+    }
+    
+    func getPoorConnectionCount() -> Int? {
+        let item = eventAttributes[EventAttributeName.poorConnectionCount.description]
+        return item??.value as? Int
+    }
+    
+    func getRetryCount() -> Int? {
+        let item = eventAttributes[EventAttributeName.retryCount.description]
+        return item??.value as? Int
+    }
+    
+    func getVideoInputErrorMessage() -> String? {
+        let item = eventAttributes[EventAttributeName.videoInputError.description]
+        return item??.value as? String
+    }
 }